WIP adding IGM model flexibility
