+++ py/redrock/test/test_utils.py
@@ -8,6 +8,7 @@ except Exception:
     cp_available = False
 
 from .. import utils
+from .. import igm
 
 class TestUtils(unittest.TestCase):
     """Test redrock utils.
@@ -43,10 +44,10 @@ class TestUtils(unittest.TestCase):
         x = np.arange(300)+800.
         myz = np.arange(11)*0.1
 
-        c = utils.transmission_Lyman(myz,x)
+        c = igm.transmission_Lyman(myz,x)
         self.assertTrue(type(c) == np.ndarray)
         for i in range(len(myz)):
-            a = utils.transmission_Lyman(myz[i], x)
+            a = igm.transmission_Lyman(myz[i], x)
             self.assertTrue(np.allclose(c[i], a))
             self.assertTrue(type(a) == np.ndarray)
 
@@ -57,13 +58,13 @@ class TestUtils(unittest.TestCase):
         x = np.arange(300)+800.
         myz = np.arange(11)*0.1
 
-        g = utils.transmission_Lyman(myz,x,use_gpu=True)
+        g = igm.transmission_Lyman(myz,x,use_gpu=True)
         self.assertTrue(type(g) == cp.ndarray)
-        c = utils.transmission_Lyman(myz,x)
+        c = igm.transmission_Lyman(myz,x)
         self.assertTrue(type(c) == np.ndarray)
         self.assertTrue(np.allclose(c, g))
         for i in range(len(myz)):
-            a = utils.transmission_Lyman(myz[i], x)
+            a = igm.transmission_Lyman(myz[i], x)
             self.assertTrue(np.allclose(g[i], a))
             self.assertTrue(type(a) == np.ndarray)
